refactor: split capital and funding modules
