+++ idl/omegax_protocol.source-hash
@@ -1,14 +1,25 @@
-c3c9d88ca82a4436ac7718712788d2984ecda68d80616991b7a8e4bff3af9f98
+9e9afb68b64188584f05133b8668035d357fc190138f9a08cdcca124de25a463
   Cargo.toml
   programs/omegax_protocol/Cargo.toml
   programs/omegax_protocol/src/args.rs
-  programs/omegax_protocol/src/capital.rs
+  programs/omegax_protocol/src/capital/allocations.rs
+  programs/omegax_protocol/src/capital/classes.rs
+  programs/omegax_protocol/src/capital/impairments.rs
+  programs/omegax_protocol/src/capital/liquidity_pool.rs
+  programs/omegax_protocol/src/capital/lp_positions.rs
+  programs/omegax_protocol/src/capital/mod.rs
+  programs/omegax_protocol/src/capital/redemptions.rs
   programs/omegax_protocol/src/claims.rs
   programs/omegax_protocol/src/constants.rs
   programs/omegax_protocol/src/errors.rs
   programs/omegax_protocol/src/events.rs
   programs/omegax_protocol/src/fees.rs
-  programs/omegax_protocol/src/funding_obligations.rs
+  programs/omegax_protocol/src/funding_obligations/funding_lines.rs
+  programs/omegax_protocol/src/funding_obligations/inflows.rs
+  programs/omegax_protocol/src/funding_obligations/mod.rs
+  programs/omegax_protocol/src/funding_obligations/obligations.rs
+  programs/omegax_protocol/src/funding_obligations/reserves.rs
+  programs/omegax_protocol/src/funding_obligations/settlement.rs
   programs/omegax_protocol/src/governance.rs
   programs/omegax_protocol/src/kernel.rs
   programs/omegax_protocol/src/kernel/auth.rs
